fix: remove circular dependency warnings when generating bundled files with Rollup
