+++ packages/@lightningjs/ui-components/src/components/TextBox/TextBox.style.d.ts
@@ -1,0 +1,9 @@
+import lng from '@lightningjs/core';
+
+type TextBoxStyle = {
+  offsetY: number;
+  offsetX: number;
+  textStyle: lng.textures.TextTexture.Settings | string;
+};
+
+export { TextBoxStyle as default }
