+++ packages/@lightningjs/ui-components/src/components/Tooltip/Tooltip.d.ts
@@ -19,7 +19,8 @@
 import lng from '@lightningjs/core';
 import Base from '../Base';
 import { Color, StylePartial } from '../../types/lui';
-import TextBox, { TextBoxStyle } from '../TextBox';
+import TextBox from '../TextBox';
+import type TextBoxStyle from '../TextBox/TextBox.style';
 
 type TooltipStyle = {
   backgroundColor: Color;
