+++ packages/@lightningjs/ui-components/src/components/index.d.ts
@@ -85,12 +85,15 @@ export { default as ProgressBar, ProgressBarStyle } from './ProgressBar';
 export { default as Provider, ProviderStyle } from './Provider';
 export { default as Radio, RadioStyle } from './Radio';
 export { default as Row } from './Row';
-export { default as ScrollWrapper, ScrollWrapperStyle } from './ScrollWrapper';
+export { default as ScrollWrapper } from './ScrollWrapper';
+export { default as ScrollWrapperStyle } from './ScrollWrapper/ScrollWrapper.style';
 export { default as Shadow, ShadowStyle } from './Shadow';
 export { default as Slider, SliderStyle } from './Slider';
 export { default as Surface, SurfaceStyle } from './Surface';
 export { default as TabBar, TabBarStyle, Tab, TabStyle } from './TabBar';
-export { default as TextBox, TextBoxStyle, TextContent } from './TextBox';
+export { default as TextBox } from './TextBox';
+export { default as TextBoxStyle } from './TextBox/TextBox.style';
+export { default as TextContent } from './TextBox/TextBox.content';
 export { default as Tile, TileStyle } from './Tile';
 export { default as TitleRow, TitleRowStyle } from './TitleRow';
 export { default as Toggle, ToggleStyle } from './Toggle';
